* org-citeproc.hs: cleanup to avoid compiler warnings   - remove unneeded imports: Data.Map, Data.Maybe   - provide showJSON methods for CitationsData, CitationData   - don't shadow bindings for `citations' and `cites'   - provide defaults when constructing WriterOptions
There is one remaining warning, that the JSON instance for Cite is an
orphan instance (since neither JSON nor Cite is defined here).  I
don't see a way to avoid this without wrapping Cite in an unnecessary
type defined in this module.  This may get fixed in the future, if
wrapping Cite becomes necessary to deal with things like common
prefixes and suffixes on multi-cites.